Switch which end of the deque subscriptions push and pop from.
Since it is a deque, it doesn't *really* matter whether
we push from the back and pull from the front, or push
from the front and pull from the back.  Switch to pushing
to the back and pulling from the front so we match what
the clients and services are doing.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>