More overall improvements to the code (#98)
* Move more implementation into rmw_subscription_data_t.

This moves more of the implementation out of rmw_zenoh.cpp
and into the class handling the subscription data.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Move more of the implementation into rmw_service_data_t.

This just allows us to hide more of the implementation and
remove it from rmw_zenoh.cpp.  It also fixes some of the
locking, which wasn't correct previously.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Move more of the implementation into rmw_client_data_t.

This allows us to encapsulate more of the client data in
the right place, and also allows us to fix the locking.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Make sure to reset the trigger.

That is, once we have discovered what the guard condition
status is, we should immediately reset the trigger condition
and wait for the next one.  Otherwise we risk triggering
over and over again.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

---------

Signed-off-by: Chris Lalancette <clalancette@gmail.com>