Fix race conditions in rmw_wait and map queries to clients (#153)
* Fix a race condition in rmw_wait.

To very briefly explain, rmw_wait:
1.  Checks to see if any of the entities (subscriptions, clients, etc)
have data ready to go.
2.  If they have data ready to go, then we skip attaching the
condition variable and waiting.
3.  If they do not have data ready to go, then we attach the
condition variable to all entities, take the condition variable
lock, and call wait_for/wait on the condition variable.
4.  Regardless of whether we did 3 or 4, we check every entity
to see if there is data ready, and mark that as appropriate
in the wait set.

There is a race in all of this, however.  If data comes in after
we've checked the entity (1), but before we've attached the
condition variable (3), then we will never be woken up.  In most
cases, this means that we'll wait the full timeout for the wait_for,
which is not what we want.

Fix this by adding another step to 3.  In particular, after we've
locked the condition variable mutex, check the entities again.
Since we change the entities to *also* take the lock before we
notify, this ensures that the entities cannot make changes that
get lost.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Small update to a comment.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Don't return an error if we can't find a number in the sequence map.

I'm not really sure that this is correct, but do it for now.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Fix query queue for multiple clients.

In particular, make sure that we track requests from
individual clients separately so that we don't mix them
up.  To do that, we store the client gid in the server
set along with the sequence_number and Query itself.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Finish changes

Signed-off-by: Yadunund <yadunund@openrobotics.org>

* Tweak api to store and retrieve query

Signed-off-by: Yadunund <yadunund@openrobotics.org>

* Lint

Signed-off-by: Yadunund <yadunund@openrobotics.org>

---------

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
Signed-off-by: Yadunund <yadunund@openrobotics.org>
Co-authored-by: Chris Lalancette <clalancette@gmail.com>