Fix compilation when building zenohc in debug mode.
When zenohc is built in debug mode, it actually changes
the name of the zenohcConfig.cmake file to zenohc_debugConfig.cmake.
Luckily it does not change the name of the library.

To deal with this, first quietly look for the debug version.
If we don't find that, then fall back to the release version.
This should make it compile in both circumstances.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>