Fix possible race in rmw_wait.
With the current code in rmw_wait, there is a race between
the time when we check whether we should wait, and whether
an event comes in.  If we've determined we should wait,
and an event comes in (and does cv.notify()) *before* we
start waiting, then we won't wake up until the timeout.

This commit fixes that by taking a mutex across *all* of
rmw_wait().  This means that if an event comes in while
we are holding the mutex, it will actually cause us to
never go to sleep.  This fixes the race.

While we are in here, we notice that we can improve performance
by just having a single lock in the context that everything
uses.  This allows us to skip the entire attach/detach dance,
getting rid of one of the 3 loops in rmw_wait.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>