Use an ordered_map to store the TopicMap.
Some of the tests in ROS 2 assume that the order in which
entities (pub, sub, clients, services) are added to the system
is the same order in which they will be returned from
calls like get_topic_names_and_types().  To facilitate that,
use an ordered_map for the mapping between a topic name and
the topic type, which remembers insertion order.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>