Implement rmw_test_fixture to start the Zenoh router (#583)
* WIP: Implent rmw_test_fixture to start the Zenoh router

Signed-off-by: Scott K Logan <logans@cottsay.net>

* Proper arbitrary port allocation for the test fixture

It's unfortunate that we need to take a dependency to parse the JSON
string coming out of the session, but this strategy should reliably open
start the router on an arbitrary unused port for the isolated tests to
use.

* Stop connecting to an upstream router

If the Zenoh config specifies a connect/endpoints value, ignore it
during isolated testing. The user can disable isolation if they actually
want the tests to connect to a specific endpoint.

* PR feedback: error handling, includes, etc

Signed-off-by: Scott K Logan <logans@cottsay.net>

* Explicitly disable multicast discovery during testing

* Set missing DLL export flag for Windows

---------

Signed-off-by: Scott K Logan <logans@cottsay.net>